+++ Shop.Domain/Aggregates/AccountAggregate/Events/AccountCreated.cs
@@ -5,13 +5,13 @@ namespace Shop.Domain.Aggregates.AccountAggregate.Events
 {
     public class AccountCreated : DomainEvent
     {
-        public AccountCreated(Guid sourceId, Guid userId, int accountNumber) : base(sourceId)
+        public AccountCreated(Guid sourceId, Guid userId, long accountNumber) : base(sourceId)
         {
             AccountNumber = accountNumber;
             UserId = userId;
         }
 
-        public int AccountNumber { get; }
+        public long AccountNumber { get; }
         public Guid UserId { get; }
     }
 }
